Updating FrequencyAnalyser
Moved filterMostFrequent and formatMap to seperate classes called
FilterMostFrequent and MapFormatter
Removed useless code
