+++ src/com/clouway/FrequencyAnalyser.java
@@ -18,26 +18,4 @@ public class FrequencyAnalyser {
         }
         return dataMap;
     }
-
-    public String formatMap(String word, ListForAnalysisDataSource dataSource){
-        String output = "";
-        for (Map.Entry<String,Integer> entry : frequencyAnalysis(word,dataSource).entrySet()){
-            output += entry.getKey() + ":";
-            output += entry.getValue() + ", ";
-        }
-        return output;
-    }
-
-    public String filterMostFrequent(String word, ListForAnalysisDataSource dataSource){
-        Map<String,Integer> data = frequencyAnalysis(word,dataSource);
-        int maxValue = Collections.max(data.values());
-        StringBuilder output = new StringBuilder();
-
-        for(Map.Entry<String,Integer> entry : data.entrySet()) {
-            if(entry.getValue()==maxValue){
-                output.append(entry.getKey()).append(" ");
-            }
-        }
-        return output.toString();
-    }
 }
